requirements.txt
openai>=1.0.0
pdf2image>=1.16.0
python-dotenv>=1.0.0
tqdm>=4.65.0
Pillow>=10.0.0
